requirements.txt
fastapi
uvicorn[standard]
aiosqlite
apscheduler
httpx
jinja2
python-dotenv
